Sat Code

+++ lib/controllers/cart_controller.dart
@@ -5,21 +5,31 @@ class CartController {
   final SupabaseClient _client = Supabase.instance.client;
   final String table = CartFields.table;
 
+  // CREATE
   Future<void> addCartItem(CartItemModel item) async {
-    await _client.from(table).insert(item.toMap());
+    try {
+      await _client.from(table).insert(item.toMap());
+    } catch (e) {
+      print('Error adding cart item: $e');
+      rethrow;
+    }
   }
 
+  // READ
   Future<List<CartItemModel>> fetchCartItems({String? userId}) async {
     final response = await _client
         .from(table)
         .select()
         .eq(CartFields.userId, userId ?? '');
 
+    print(response);
+
     return (response as List)
         .map((item) => CartItemModel.fromMap(item as Map<String, dynamic>))
         .toList();
   }
 
+  // READ
   Future<CartItemModel?> getCartItemById(String id) async {
     final response = await _client
         .from(table)
@@ -33,22 +43,36 @@ class CartController {
     return null;
   }
 
+  // UPDATE
   Future<void> updateCartItem(CartItemModel item) async {
-    await _client
-        .from(table)
-        .update(item.toMap())
-        .eq(CartFields.id, item.id as String);
+    try {
+      await _client
+          .from(table)
+          .update(item.toMap())
+          .eq(CartFields.id, item.id.toString());
+      print('Updated: ${item.toMap()}');
+    } catch (e) {
+      print('Error updating cart item: $e');
+      rethrow;
+    }
   }
 
+  // DELETE
   Future<void> deleteCartItem(String id) async {
-    await _client.from(table).delete().eq(CartFields.id, id);
+    try {
+      await _client.from(table).delete().eq(CartFields.id, id);
+    } catch (e) {
+      print('Error deleting cart item: $e');
+      rethrow;
+    }
   }
 
   Future<void> clearCart({required String userId}) async {
-    await _client
-        .from(table)
-        .delete()
-        .eq(CartFields.userId, userId);
+    try {
+      await _client.from(table).delete().eq(CartFields.userId, userId);
+    } catch (e) {
+      print('Error clearing cart: $e');
+      rethrow;
+    }
   }
-
 }
